Add files via upload
This repository contains a MATLAB implementation of a program that can embed a secret image into a cover image using Least Significant Bit (LSB) steganography and Advanced Encryption Standard (AES) encryption. The user can enter an encryption key, which is used to encrypt the secret image before it is embedded into the cover image. The resulting encoded image can be saved to a file.

The program also includes a function to extract and decode the secret image from the encoded image, using LSB steganography and AES decryption. The user must enter the same encryption key that was used to embed the secret image in order to successfully extract and decode the secret image. The decoded secret image can be saved to a file.

This code provides a way to securely embed and extract a secret image within a cover image, which can be useful for applications such as digital watermarking, data hiding, and secure communications.

To use this code, simply run the main.m file and follow the instructions provided. You will be prompted to enter the encryption key and the filenames of the cover and secret images. The program will then embed the secret image into the cover image, and save the resulting encoded image to a file. To extract and decode the secret image, run the main.m file again and choose the 'decode' option. You will be prompted to enter the encryption key and the filename of the encoded image. The program will then extract and decode the secret image, and save it to a file.